Add FunnelSeries component
