+++ src/FunnelChart/FunnelSeries/FunnelAxis/index.ts
