+++ src/FunnelChart/FunnelSeries/index.tsx
@@ -1,0 +1,3 @@
+export * from './FunnelSeries';
+export * from './FunnelArc';
+export * from './FunnelAxis';
