+++ src/FunnelChart/index.ts
@@ -1,3 +1,2 @@
 export * from './FunnelChart';
-export * from './FunnelArc';
-export * from './FunnelAxis';
+export * from './FunnelSeries';
